Refactor main `useEffect` to use hooks.
