+++ avivator/src/Avivator.js
@@ -1,34 +1,21 @@
-import React, { useState, useEffect, useReducer } from 'react';
-import Button from '@material-ui/core/Button';
+import React, { useEffect } from 'react';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import Grid from '@material-ui/core/Grid';
 import Snackbar from '@material-ui/core/Snackbar';
 import Alert from '@material-ui/lab/Alert';
 
-import AddIcon from '@material-ui/icons/Add';
-
 import {
   SideBySideViewer,
   PictureInPictureViewer,
-  VolumeViewer,
-  getChannelStats
+  VolumeViewer
 } from '../../dist';
-import {
-  createLoader,
-  channelsReducer,
-  useWindowSize,
-  buildDefaultSelection,
-  getNameFromUrl,
-  getSingleSelectionStats,
-  guessRgb,
-  range
-} from './utils';
+import { useWindowSize, getNameFromUrl, guessRgb } from './utils';
 import {
   useChannelSettings,
-  useChannelSetters,
   useImageSettingsStore,
   useViewerStore
 } from './state';
+import { initImage, init3DSettings } from './hooks';
 
 import ChannelController from './components/ChannelController';
 import Menu from './components/Menu';
@@ -50,13 +37,7 @@ import {
 } from './components/SnackbarAlerts';
 import { DropzoneWrapper } from './components/Dropzone';
 
-import {
-  MAX_CHANNELS,
-  DEFAULT_OVERVIEW,
-  FILL_PIXEL_VALUE,
-  GLOBAL_SLIDER_DIMENSION_FIELDS,
-  COLOR_PALLETE
-} from './constants';
+import { DEFAULT_OVERVIEW, GLOBAL_SLIDER_DIMENSION_FIELDS } from './constants';
 import './index.css';
 
 /**
@@ -71,9 +52,6 @@ export default function Avivator(props) {
 
   const viewSize = useWindowSize();
 
-  const [metadata, setMetadata] = useState(null);
-  const [source, setSource] = useState(initSource);
-  const [dimensions, setDimensions] = useState([]);
   const {
     colors,
     sliders,
@@ -82,13 +60,6 @@ export default function Avivator(props) {
     selections,
     loader
   } = useChannelSettings();
-  const {
-    setPropertiesForChannels,
-    setPropertiesForChannel,
-    addChannels,
-    addChannel,
-    setLoader
-  } = useChannelSetters();
   const {
     lensSelection,
     colormap,
@@ -111,197 +82,17 @@ export default function Avivator(props) {
     use3d,
     toggleIsOffsetsSnackbarOn,
     toggleIsNoImageUrlSnackbarOn,
-    toggleUseLinkedView,
-    toggleIsOverviewOn,
-    toggleZoomLock,
-    toggleUse3d,
-    setViewerState
+    useColormap,
+    setViewerState,
+    channelOptions,
+    metadata,
+    source,
+    useLens
   } = useViewerStore();
+  useEffect(() => setViewerState('source', initSource), []);
 
-  useEffect(() => {
-    async function changeLoader() {
-      setViewerState('isLoading', true);
-      const { urlOrFile } = source;
-      const {
-        data: nextLoader,
-        metadata: nextMeta
-      } = await createLoader(urlOrFile, toggleIsOffsetsSnackbarOn, message =>
-        setViewerState('loaderErrorSnackbar', { on: true, message })
-      );
-
-      if (nextLoader) {
-        const newSelections = buildDefaultSelection(nextLoader[0]);
-        const { Channels } = nextMeta.Pixels;
-        const channelOptions = Channels.map((c, i) => c.Name ?? 'Channel ' + i);
-        // Default RGB.
-        let newSliders = [
-          [0, 255],
-          [0, 255],
-          [0, 255]
-        ];
-        let newDomains = [
-          [0, 255],
-          [0, 255],
-          [0, 255]
-        ];
-        let newColors = [
-          [255, 0, 0],
-          [0, 255, 0],
-          [0, 0, 255]
-        ];
-        const lowResSource = nextLoader[nextLoader.length - 1];
-        const isRgb = guessRgb(nextMeta);
-        if (!isRgb) {
-          const stats = await Promise.all(
-            newSelections.map(async selection => {
-              const raster = await lowResSource.getRaster({ selection });
-              return getChannelStats(raster.data);
-            })
-          );
-
-          newDomains = stats.map(stat => stat.domain);
-          newSliders = stats.map(stat => stat.autoSliders);
-          // If there is only one channel, use white.
-          newColors =
-            stats.length === 1
-              ? [[255, 255, 255]]
-              : stats.map((_, i) => COLOR_PALLETE[i]);
-        } else if (isRgb || channelOptions.length === 1) {
-          // RGB should not use a lens.
-          isLensOn && toggleIsLensOn(); // eslint-disable-line no-unused-expressions
-        }
-
-        const { labels, shape } = lowResSource;
-        const newDimensions = labels.map((l, i) => {
-          return {
-            field: l,
-            values: l === 'c' ? channelOptions : range(shape[i])
-          };
-        });
-        setDimensions(newDimensions);
-        addChannels(
-          ['ids', 'selections', 'domains', 'sliders', 'colors'],
-          [
-            newDomains.map(() => String(Math.random())),
-            newSelections,
-            newDomains,
-            newSliders,
-            newColors
-          ]
-        );
-        setLoader(nextLoader);
-        setMetadata(nextMeta);
-        setViewerState('isLoading', false);
-        setViewerState(
-          'pixelValues',
-          new Array(newSelections.length).fill(FILL_PIXEL_VALUE)
-        );
-        // Set the global selections (needed for the UI). All selections have the same global selection.
-        setViewerState('globalSelection', newSelections[0]);
-        if (use3d) toggleUse3d();
-        // eslint-disable-next-line no-unused-expressions
-        history?.push(
-          typeof urlOrFile === 'string' ? `?image_url=${urlOrFile}` : ''
-        );
-      }
-    }
-    changeLoader();
-  }, [source, history]); // eslint-disable-line react-hooks/exhaustive-deps
-
-  useEffect(() => {
-    async function updateStatsFor3D() {
-      if (!use3d) {
-        const lowResSource = loader[loader.length - 1];
-        const stats = await Promise.all(
-          selections.map(async selection => {
-            const raster = await lowResSource.getRaster({ selection });
-            return getChannelStats(raster.data);
-          })
-        );
-        const domains = stats.map(stat => stat.domain);
-        const sliders = stats.map(stat => stat.autoSliders);
-
-        domains.forEach((domain, index) =>
-          dispatch({
-            type: 'CHANGE_DOMAIN',
-            value: domain,
-            index
-          })
-        );
-        sliders.forEach((slider, index) =>
-          dispatch({
-            type: 'CHANGE_SLIDER',
-            value: slider,
-            index
-          })
-        );
-      } else {
-        const lowResSource = loader[loader.length - 1];
-        const { labels, shape } = lowResSource;
-        const sizeZ = shape[labels.indexOf('z')] >> (loader.length - 1);
-        const stats = await Promise.all(
-          selections.map(async selection => {
-            const raster0 = await lowResSource.getRaster({
-              selection: { ...selection, z: 0 }
-            });
-            const rasterMid = await lowResSource.getRaster({
-              selection: { ...selection, z: Math.floor(sizeZ / 2) }
-            });
-            const rasterTop = await lowResSource.getRaster({
-              selection: { ...selection, z: sizeZ - 1 }
-            });
-            const stats0 = getChannelStats(raster0.data);
-            const statsMid = getChannelStats(rasterMid.data);
-            const statsTop = getChannelStats(rasterTop.data);
-            return {
-              domain: [
-                Math.min(
-                  stats0.domain[0],
-                  statsMid.domain[0],
-                  statsTop.domain[0]
-                ),
-                Math.max(
-                  stats0.domain[1],
-                  statsMid.domain[1],
-                  statsTop.domain[1]
-                )
-              ],
-              autoSliders: [
-                Math.min(
-                  stats0.autoSliders[0],
-                  statsMid.autoSliders[0],
-                  statsTop.autoSliders[0]
-                ),
-                Math.max(
-                  stats0.autoSliders[1],
-                  statsMid.autoSliders[1],
-                  statsTop.autoSliders[1]
-                )
-              ]
-            };
-          })
-        );
-        const domains = stats.map(stat => stat.domain);
-        const sliders = stats.map(stat => stat.autoSliders);
-
-        domains.forEach((domain, index) =>
-          dispatch({
-            type: 'CHANGE_DOMAIN',
-            value: domain,
-            index
-          })
-        );
-        sliders.forEach((slider, index) =>
-          dispatch({
-            type: 'CHANGE_SLIDER',
-            value: slider,
-            index
-          })
-        );
-      }
-    }
-    updateStatsFor3D();
-  }, [use3d, source]); // eslint-disable-line react-hooks/exhaustive-deps
+  initImage(source, history);
+  init3DSettings();
 
   const handleSubmitNewUrl = (event, url) => {
     event.preventDefault();
@@ -310,7 +101,7 @@ export default function Avivator(props) {
       // Use the trailing part of the URL (file name, presumably) as the description.
       description: getNameFromUrl(url)
     };
-    setSource(newSource);
+    setViewerState('source', newSource);
   };
   const handleSubmitFile = files => {
     let newSource;
@@ -326,13 +117,14 @@ export default function Avivator(props) {
         description: 'data.zarr'
       };
     }
-    setSource(newSource);
+    setViewerState('source', newSource);
   };
   const isRgb = metadata && guessRgb(metadata);
-  const globalControlDimensions = dimensions?.filter(dimension =>
-    GLOBAL_SLIDER_DIMENSION_FIELDS.includes(dimension.field)
-  );
-  const channelOptions = dimensions.filter(j => j.field === 'c')[0]?.values;
+  const globalControlDimensions =
+    loader[0] &&
+    loader[0].labels?.filter(dimension =>
+      GLOBAL_SLIDER_DIMENSION_FIELDS.includes(dimension.field)
+    );
   const channelControllers = ids.map((id, i) => {
     const name = channelOptions[selections[i].c];
     return (
@@ -351,12 +143,14 @@ export default function Avivator(props) {
       </Grid>
     );
   });
-  const globalControllers = globalControlDimensions.map(dimension => {
-    // Only return a slider if there is a "stack."
-    return dimension.values.length > 1 && !use3d ? (
-      <GlobalSelectionSlider key={dimension.field} dimension={dimension} />
-    ) : null;
-  });
+  const globalControllers =
+    globalControlDimensions &&
+    globalControlDimensions.map(dimension => {
+      // Only return a slider if there is a "stack."
+      return dimension.values.length > 1 && !use3d ? (
+        <GlobalSelectionSlider key={dimension.field} dimension={dimension} />
+      ) : null;
+    });
   return (
     <>
       {
@@ -424,9 +218,9 @@ export default function Avivator(props) {
           urlOrFile={source.urlOrFile}
           handleSubmitFile={handleSubmitFile}
         >
-          {!isRgb && <ColormapSelect />}
+          {useColormap && <ColormapSelect />}
           {use3d && <RenderingModeSelect />}
-          {!isRgb && channelOptions?.length > 1 && !colormap && !use3d && (
+          {useLens && !colormap && (
             <LensSelect
               channelOptions={selections.map(sel => channelOptions[sel.c])}
             />
